requirements.txt
python-dotenv
ipykernel
pypdf
pymupdf
bs4
arxiv
wikipedia
langchain
langchain-community
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
streamlit
tiktoken
watchdog
duckduckgo-search
streamlit